+++ net/http/Service.php
@@ -79,7 +79,7 @@ class Service extends \lithium\core\Object {
 	 *        - `'username'` _string_
 	 *        - `'password'` _string_
 	 *        - `'encoding'` _string_
-	 *        - `'socket'` _string_
+	 *        - `'socket'` _string_|_boolean_ if `false` will not establish a connection.
 	 * @return void
 	 */
 	public function __construct(array $config = []) {
@@ -96,21 +96,13 @@ class Service extends \lithium\core\Object {
 			'socket'     => 'Context'
 		];
 		parent::__construct($config + $defaults);
-	}
 
-	/**
-	 * Initialize connection.
-	 *
-	 * @return void
-	 */
-	protected function _init() {
-		$config = ['classes' => $this->_classes] + $this->_config;
-
-		try {
-			$this->connection = Libraries::instance('socket', $config['socket'], $config);
-		} catch(ClassNotFoundException $e) {
-			$this->connection = null;
+		if ($this->_config['socket']) {
+			$this->connection = Libraries::instance(
+				'socket', $this->_config['socket'], ['classes' => $this->_classes] + $this->_config
+			);
 		}
+
 		$this->_responseTypes += [
 			'headers' => function($response) { return $response->headers; },
 			'body' => function($response) { return $response->body(); },